add shutdown guard to profiler tests
