+++ src/tests/profiler/native/profiler.cpp
@@ -3,6 +3,9 @@
 
 #include "profiler.h"
 
+std::atomic<bool> ShutdownGuard::s_preventHooks;
+std::atomic<int> ShutdownGuard::s_hooksInProgress;
+
 Profiler::Profiler() : refCount(0), pCorProfilerInfo(nullptr)
 {
 }
@@ -18,6 +21,8 @@ Profiler::~Profiler()
 
 HRESULT STDMETHODCALLTYPE Profiler::Initialize(IUnknown *pICorProfilerInfoUnk)
 {
+    ShutdownGuard::Initialize();
+
     printf("Profiler.dll!Profiler::Initialize\n");
     fflush(stdout);
 
@@ -35,6 +40,9 @@ HRESULT STDMETHODCALLTYPE Profiler::Shutdown()
 {
     printf("Profiler.dll!Profiler::Shutdown\n");
     fflush(stdout);
+
+    // Wait for any in progress profiler callbacks to finish.
+    ShutdownGuard::WaitForInProgressHooks();
 
     if (this->pCorProfilerInfo != nullptr)
     {
