+++ src/tests/profiler/native/profiler.h
@@ -7,6 +7,8 @@
 
 #include <atomic>
 #include <cstdio>
+#include <thread>
+#include <chrono>
 #include "cor.h"
 #include "corprof.h"
 #include "holder.h"
@@ -22,6 +24,68 @@
 #define SHORT_LENGTH    32
 #define STRING_LENGTH  256
 #define LONG_LENGTH   1024
+
+class ShutdownGuard
+{
+private:
+    static std::atomic<bool> s_preventHooks;
+    static std::atomic<int> s_hooksInProgress;
+
+public:
+    ShutdownGuard()
+    {
+        ++s_hooksInProgress;
+    }
+
+    ~ShutdownGuard()
+    {
+        --s_hooksInProgress;
+    }
+
+    static void Initialize()
+    {
+        s_preventHooks = false;
+        s_hooksInProgress = 0;
+    }
+
+    static bool HasShutdownStarted()
+    {
+        return s_preventHooks.load();
+    }
+
+    static void WaitForInProgressHooks()
+    {
+        s_preventHooks = true;
+
+        while (s_hooksInProgress.load() > 0)
+        {
+            std::this_thread::sleep_for(std::chrono::milliseconds(10));
+        }
+    }
+};
+
+// Managed code can keep running after Shutdown() is called, and things like
+// ELT hooks will continue to be called. We would AV if we tried to call
+// in to freed resources.
+#define SHUTDOWNGUARD()                             \
+    do                                              \
+    {                                               \
+        ShutdownGuard();                            \
+        if (ShutdownGuard::HasShutdownStarted())    \
+        {                                           \
+            return S_OK;                            \
+        }                                           \
+    } while(0) 
+
+#define SHUTDOWNGUARD_RETVOID()                     \
+    do                                              \
+    {                                               \
+        ShutdownGuard();                            \
+        if (ShutdownGuard::HasShutdownStarted())    \
+        {                                           \
+            return;                                 \
+        }                                           \
+    } while(0)
 
 class Profiler : public ICorProfilerCallback10
 {
